requirements.txt
selenium
flask==2.0.1
Werkzeug==2.0.1
flask-socketio==5.1.1
uvicorn[standard]
gevent-websocket==0.10.1
gevent==21.12.0
eventlet==0.33.3
python-dotenv
python-engineio==4.2.1
python-socketio==5.4.0
mysql-connector-python
nats-py
asyncio
pytest==7.3.1
pytest-flask==1.2.0
flake8==6.0.0
mypy==1.3.0
asgiref==3.4.1
uvicorn[standard]
pytest-asyncio==0.21.0